move: 모노레포 구조 1차 변경 (#24)
