+++ apps/web/app/layout.tsx
@@ -1,8 +1,8 @@
 import QueryProvider from '../providers/QueryProvider';
-import PageContainer from '@repo/ui/components/layout/PageContainer';
+import PageContainer from '../components/layout/PageContainer';
 import { home } from '../metadata/common/home';
 import { pretendard } from '../fonts';
-import TabBar from '@repo/ui/components/common/TabBar';
+import TabBar from '../components/common/TabBar';
 import './globals.css';
 
 export const metadata = home;
