+++ apps/web/components/auctions/detail/AuctionDetailCard.tsx
