+++ apps/web/lib/supabase/client/client.ts
@@ -1,5 +1,5 @@
-import { Database } from '@repo/ui/types/supabase';
 import { createBrowserClient } from '@supabase/ssr';
+import { Database } from '../../../types/supabase';
 
 export function createClient() {
   return createBrowserClient<Database>(
