+++ apps/web/lib/supabase/query/auctions.ts
