+++ apps/web/utils/formatNumber.ts
