--- packages/ui/src/components/card.tsx
@@ -1,20 +1,0 @@
-import { type ReactNode } from 'react';
-
-export function Card({ title, children, href }: { title: string; children: ReactNode; href: string }) {
-  return (
-    <a
-      className="group rounded-lg border border-transparent px-5 py-4 transition-colors hover:border-neutral-700 hover:bg-neutral-800/30"
-      href={`${href}?utm_source=create-turbo&utm_medium=with-tailwind&utm_campaign=create-turbo"`}
-      rel="noopener noreferrer"
-      target="_blank"
-    >
-      <h2 className="mb-3 text-2xl font-semibold">
-        {title}{' '}
-        <span className="inline-block transition-transform group-hover:translate-x-1 motion-reduce:transform-none">
-          -&gt;
-        </span>
-      </h2>
-      <p className="m-0 max-w-[30ch] text-sm opacity-50">{children}</p>
-    </a>
-  );
-}
