+++ packages/ui/tsconfig.json
@@ -7,6 +7,6 @@
       "@repo/ui/*": ["./src/*"]
     }
   },
-  "include": ["src"],
+  "include": ["src", "../../apps/web/components/common", "../../apps/web/components/layout"],
   "exclude": ["dist", "build", "node_modules"]
 }
